+++ CHANGELOG.md
@@ -25,7 +25,7 @@
 * Adding `xan separate --prefix`.
 * Adding `xan heatmap -C` short flag for `--cram`.
 * Adding `xan heatmap --repeat-headers`.
-* Adding `rank` window function.
+* Adding `rank` & `cume_dist` window functions.
 
 *Quality of Life*